The company's read side is its own file
company.go held the vocabulary, the write and the assembly that reads the
record back, and crossed the 500-line ceiling. readCompany moves out whole:
one assembly called by the form's GET and by every write that answers with the
record it just saved, so what a save returns and what a later read returns
cannot drift apart.
